[tests] Refactoring
